Add option not to throw malformed_packet exception
By default, a malformed_packet exception is thrown when certain errors occur
while constructing a packet. When this occurs during construction, the entire
stack of PDUs is ripped up. Now, this can be turned off, so that the packet will
be constructed as much as possible but a malformed flag will be set. The
partially constructed PDUs will be returned to the caller.

issue: upstream #269