Add const methods to IP and TCP to calculate correct checksum
The checksum was only calculated in the write_serialization methods, but
other header fields were updated such as header length and protocol. The
new calculate_checksum methods are const, so they do not modify any
fields in the header. They are useful for checking if the checksum in
a received packet is valid or not.