Support IP/UDP checksum overriding for RoCEv2 and its icrc calculation (#7)
Add support in libtins for setting the IP and UDP checksums, and once doing so, disabling re-calculation of the checksum on serialization. This allows protocols such as RoCE v2 to set the UDP checksum to 0 as required, it also allows for calculation of icrc which needs both IP and UDP checksum to be all 1s.